feat(mithril): public Sync API
